Refactor out XmlAdaptedAccount to individual class
