+++ src/main/java/seedu/address/storage/XmlSerializableAccount.java
@@ -1,7 +1,6 @@
 package seedu.address.storage;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
@@ -9,7 +8,6 @@ import javax.xml.bind.annotation.XmlRootElement;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.account.Account;
 import seedu.address.model.account.AccountList;
-import seedu.address.model.account.Role;
 
 /**
  * An Immutable Account that is serializable to XML format
@@ -69,79 +67,3 @@ public class XmlSerializableAccount {
         return accounts.equals(((XmlSerializableAccount) other).accounts);
     }
 }
-
-/**
- * JAXB-friendly version of the Person.
- */
-class XmlAdaptedAccount {
-    @XmlElement(required = true)
-    private String userName;
-    @XmlElement(required = true)
-    private String password;
-    @XmlElement(required = true)
-    private Role role;
-
-    /**
-     * Constructs an XmlAdaptedAccount.
-     * This is the no-arg constructor that is required by JAXB.
-     */
-    public XmlAdaptedAccount() {}
-
-    /**
-     * Constructs an {@code XmlAdaptedAccount} with the given account details.
-     */
-    public XmlAdaptedAccount(String userName, String password, Role role) {
-        this.userName = userName;
-        this.password = password;
-        this.role = role;
-    }
-
-    /**
-     * Converts a given Person into this class for JAXB use.
-     *
-     * @param source future changes to this will not affect the created XmlAdaptedPerson
-     */
-    public XmlAdaptedAccount(Account source) {
-        userName = source.getUserName();
-        password = source.getPassword();
-        role = source.getRole();
-    }
-
-    /**
-     * Converts this jaxb-friendly adapted account object into the model's Account object.
-     *
-     * @throws IllegalValueException if there were any data constraints violated in the adapted person
-     */
-    public Account toModelType() throws IllegalValueException {
-
-        if (userName == null) {
-            throw new IllegalValueException("Username is null.");
-        }
-        if (password == null) {
-            throw new IllegalValueException("Password is null.");
-        }
-
-        if (role == null) {
-            throw new IllegalValueException("Role is null.");
-        }
-
-        return new Account(userName, password, role);
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        if (!(other instanceof XmlAdaptedPerson)) {
-            return false;
-        }
-
-        XmlAdaptedAccount otherAccount = (XmlAdaptedAccount) other;
-        return Objects.equals(userName, otherAccount.userName)
-                && Objects.equals(password, otherAccount.password)
-                && Objects.equals(role, otherAccount.role);
-    }
-}
-
